Change default coordinate system to Cartesian
Previously most methods returned the result in spherical coordinates.
The problem with spherical coordinates is that the north and south
pole do not have well defined representations, this would lead to
problems when parameterizing the boundary close to the north pole. In
Cartesian coordinates all points on the sphere have a well defined
representation, avoiding this problem.