Add orientation_raw, update eigenfunction constructors
Add the method orientation_raw and update the orientation method. They
now work similar to angle_raw and angle.

The constructors for StandaloneVertexEigenfunction and
StandaloneLightningEigenfunction are updated to make use of these new
methods.

Finally StandaloneVertexEigenfunction has been made parametric in the
types, similar to how StandaloneLightningEigenfunction was. It doesn't
use this parametric type yet though.