Rework Polygon
It now takes the vertices as the first argument instead of the angles.
The angles can then either be given directly or for non-rational
angles it can also compute them from the vertices. In addition to this
it now also stores the orientation of the first vertex, which for
non-rational angles has to be given explicitly, allowing for accurate
computations of the orientation even in the case of rational angles.